requirements.txt.txt
streamlit>=1.22.0
pandas>=1.5.0
requests>=2.28.0
matplotlib>=3.6.0
Pillow>=9.3.0